refactor: put some of the main PowerUp operations into the PowerUps repo
