+++ SuperNodes.Tests/tests/ChickensoftGeneratorTest.cs
@@ -1,11 +1,5 @@
 namespace SuperNodes.Tests;
 
-using System.Collections.Immutable;
-using System.Linq;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Moq;
 using Shouldly;
 using Xunit;
 
@@ -50,54 +44,5 @@ public class ChickensoftGeneratorTest {
     ChickensoftGenerator.If(false, "a", "b").ShouldBe("b");
     ChickensoftGenerator.If(true, "a").ShouldBe("a");
     ChickensoftGenerator.If(false, "a").ShouldBe("");
-  }
-
-  [Fact]
-  public void GetBaseClassHierarchy() {
-    var sym = new Mock<INamedTypeSymbol>();
-    var baseType = new Mock<INamedTypeSymbol>();
-    var baseBaseType = new Mock<INamedTypeSymbol>();
-
-    sym.Setup(s => s.BaseType).Returns(baseType.Object);
-    baseType.Setup(s => s.BaseType).Returns(baseBaseType.Object);
-    baseBaseType.Setup(s => s.BaseType).Returns((INamedTypeSymbol?)null);
-
-    baseType.Setup
-      (s => s.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat)
-    ).Returns("BaseType");
-
-    baseBaseType.Setup(
-      s => s.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat)
-    ).Returns("BaseBaseType");
-
-    ChickensoftGenerator.GetBaseClassHierarchy(sym.Object).ShouldBe(
-      new[] { "BaseType", "BaseBaseType" }
-    );
-  }
-
-  [Fact]
-  public void GetUsings() {
-    const string code = """
-    using A;
-    using B;
-    namespace Foo {
-      using C;
-      class Bar {
-      }
-    }
-    """;
-
-    var model = TestUtils.GetSemanticModel(code);
-
-    var classDecl = model.SyntaxTree.GetRoot()
-      .DescendantNodes()
-      .OfType<ClassDeclarationSyntax>()
-      .First();
-
-    var symbol = model.GetDeclaredSymbol(classDecl)!;
-
-    ChickensoftGenerator.GetUsings(symbol).ShouldBe(
-      new[] { "A", "B", "C" }.ToImmutableHashSet()
-    );
   }
 }
